+++ CHANGES.markdown
@@ -1,4 +1,5 @@
 ## Changes in 1.4.2
+ - Properly handle ctrl-c while running QuickCheck properties (#93)
  - Default to `--color=always` when `--color` is used without argument (#106)
 
 ## Changes in 1.4.1