+++ CHANGES.markdown
@@ -1,10 +1,12 @@
+## next
+ - Add `--dry-run` (#111)
+
 ## Changes in 1.4.2
  - Properly handle ctrl-c while running QuickCheck properties (#93)
  - Default to `--color=always` when `--color` is used without argument (#106)
  - treat `--qc-max-success` as an alias for `--maximum-generated-tests`
 
 ## Changes in 1.4.1
-
  - Used CPU time is now only included in the test summary if run with
    `--print-cpu-time`.